refactor: GroupChallengeReadService 기능별 서비스로 분리
